workspace page removed
